+++ src/Components/Layout/FLSettingsNavbarLayout/FLSettingNavbarLayout.interface.ts
@@ -1,0 +1,22 @@
+import { ReactNode } from "react";
+
+export interface IFLSettingsNavbarLayout {
+  testID: `header-${string}`;
+  isDrawerCloseable?: boolean;
+  title?: string;
+  userAvatarSrc?: string;
+  accountName?: string;
+  popUp?: ReactNode;
+  children?: ReactNode;
+  headbarPosition: "fixed" | "absolute" | "relative" | "static" | "sticky";
+  sections: {
+    title: string;
+    path: string;
+    icon: JSX.Element;
+    children?: {
+      title: string;
+      path: string;
+    }[];
+  }[];
+  handleLogOut?: () => void;
+}
