+++ src/Components/TMS/Organisms/index.tsx
@@ -1,2 +1,1 @@
-export * from "./FOHeader";
 export * from "./FOTMSSidebar";
