+++ test/Components/Layout/FLSettingsNavbarLayout/FLSettingsNavbarLayout.test.tsx
@@ -11,7 +11,11 @@ import { render } from "@testing-library/react";
 describe("FLSettingsNavbarLayout", () => {
   it("should match snapshot", () => {
     const { container } = render(
-      <FLSettingsNavbarLayout sections={[]} testID="header-settings" />
+      <FLSettingsNavbarLayout
+        sections={[]}
+        testID="header-settings"
+        headbarPosition="absolute"
+      />
     );
     expect(container).toMatchSnapshot();
   });
